chore: separate all api functions into separate respective controllers
