+++ src/API/controllers/index.ts
@@ -1,0 +1,4 @@
+export * from "./ads.controller";
+export * from "./balance.controller";
+export * from "./bot.controller";
+export * from "./readings.controller";
