+++ src/helpers.ts
@@ -1,61 +1,4 @@
-import { retrieveLaunchParams } from "@telegram-apps/sdk-react";
-import { ValidationResponse, CardKey, ImageModule } from "./types";
-
-export const validateInitData = async (): Promise<boolean> => {
-  const { initDataRaw } = retrieveLaunchParams();
-
-  try {
-    const response = await fetch(
-      import.meta.env.VITE_INIT_DATA_VALIDATION_URL,
-      {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: `tma ${initDataRaw}`,
-        },
-      }
-    );
-
-    const data: ValidationResponse = await response.json();
-
-    return data.success;
-  } catch (error) {
-    throw new Error(`${error}`);
-  }
-};
-
-export const getInvoiceLink = async (
-  title: string,
-  description: string,
-  amount: number,
-  coinsQty?: number
-): Promise<string | undefined> => {
-  try {
-    const invoiceTitle = coinsQty ? `${coinsQty} ${title}` : title;
-
-    const response = await fetch(
-      "https://tarot-bot-18921c9756be.herokuapp.com/create_invoice",
-      {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          title: invoiceTitle,
-          description: description,
-          amount: amount,
-          label: title,
-        }),
-      }
-    );
-
-    const invoiceLink = await response.json();
-
-    return invoiceLink;
-  } catch (error) {
-    console.error(`An error occured: ${error}`);
-  }
-};
+import { CardKey, ImageModule } from "./types";
 
 export const SPREADS = {
   CARD_OF_THE_DAY: {
